requirements.txt
python-telegram-bot
fastapi
gunicorn
uvicorn[standard]
python-dotenv
prometheus-client
prometheus-fastapi-instrumentator
psutil
requests
pymongo
influxdb3-python
croniter
cron-descriptor
jsons

black
pytest
pytest-asyncio
pytest-mock
mongomock
